+++ app/src/main/kotlin/com/pitchedapps/frost/facebook/FbConst.kt
@@ -5,6 +5,7 @@ package com.pitchedapps.frost.facebook
  */
 
 const val FACEBOOK_COM = "facebook.com"
+const val FBCDN_NET = "fbcdn.net"
 const val HTTPS_FACEBOOK_COM = "https://$FACEBOOK_COM"
 const val FB_URL_BASE = "https://m.$FACEBOOK_COM/"
 fun PROFILE_PICTURE_URL(id: Long) = "https://graph.facebook.com/$id/picture?type=large"